pyproject.toml
[tool.poetry]
name = "datn"
version = "0.1.0"
description = "Intelligent Learning Assistant with Interaction and Knowledge Testing"
authors = ["NamPV-Picon <nam03412003@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
pypdf2 = "^3.0.1"
python-docx = "^1.1.0"
python-pptx = "^0.6.23"
websockets = "^12.0"
httpx = "^0.26.0"
aiofiles = "^23.2.1"
bcrypt = "4.0.1"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
